backend/requirements.txt
alembic==1.13.1
cryptography==40.0.2
Flask==3.0.3
Flask_Cors==4.0.0
Flask_JWT_Extended==4.6.0
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
psycopg2-binary
openai==1.25.0
pandas==2.2.2
pdf2image==1.17.0
pytesseract==0.3.10
python-dotenv==1.0.1
SQLAlchemy==2.0.28
textract==1.6.5
Werkzeug==3.0.2
bcrypt
Pillow